Fix ctags check for different global vars.
[Problem]
Before submitting my last change I forgot to rebase my branch with the mainline
and this caused the ctags check to check if one variable existed and if other
totally non related one was not empty.

[Solution]
In the ctags check use the same global variable so the test can actually pass.